Change db table name and fields
Change some fields in database table and table name.
Table name change from bitcoin_button_coinbase to bitcoin_button_transaction.
Plugin is set up to do automatic update of database schema but in this case
it will not work since the name changed.
Also the backend field will be missing and should be set to "coinbase".